Hexagonal Lattice and seed-driven randomization: - Created class HexLattice to represent a hexagonal coordinate system lattice - Created methods to print various SOM  representations   (method for printing hex-lattice is floppy) - Added parameter seed to normalizedRandomInit method on Lattice and RandomVectorSet's constructor - Added a Random instance ("rand") both in the method and the class to control the generated sequence - Added variables "initSeed" and "shuffleSeed" to main to define seed for randomization

+++ src/som/FunctionCollector.scala
@@ -28,7 +28,7 @@ object FunctionCollector {
    * @param neighRadius Current neighborhood radius
    * @return
    */
-  def gaussianNeighborhood (bmuX: Int, bmuY: Int, neighX: Int, neighY: Int, neighRadius: Double): Double = {
+  def gaussianNeighborhood (bmuX: Float, bmuY: Float, neighX: Float, neighY: Float, neighRadius: Double): Double = {
     val distance = math.pow(bmuX - neighX, 2) + math.pow(bmuY - neighY, 2)
 
     if (distance == 0) 1
